+++ __tests__/asset.test.ts
@@ -15,6 +15,7 @@ function repeat( str: string, count: number ): string {
 
 test("asset::extra", () => {
     expect( Number(asset(1.123, symbol("4,EOS")).amount) ).toBe(1);
+    asset(asset("1.0000 EOS"));
 });
 
 test("asset::eosiolib", () => {